Moved displaying of titles into show functions
added tabbing to be allowed to print recursive packets properly
Made PGPMessage contain a Tag8::Ptr to fulfill RFC 4880 section 2.3
properly
fixed a few typos in main.cpp

+++ PGPCleartextSignature.cpp
@@ -117,7 +117,7 @@ void PGPCleartextSignature::read(std::ifstream & file){
 }
 
 std::string PGPCleartextSignature::show(const uint8_t indents, const uint8_t indent_size) const{
-    uint8_t tab = indents * indent_size;
+    unsigned int tab = indents * indent_size;
     return std::string(tab, ' ') + "Message:\n" + message + "\n\n" + std::string(tab, ' ') + sig.show(indents, indent_size);
 }
